BUG-10181 Fix the property name for the tserver port.
At the moment, we don't want to use the constants defined in
o.a.a.c.c.Property as this would require us to bundle the accumulo-core
jar and ensure that YARN has access to it. By hard-coding the property
name, we can avoid having to deal with this classpath issues for the
time being.